+++ src/wg_federation/data_transformation/loader/configuration_loader_interface.py
@@ -10,7 +10,7 @@ class ConfigurationLoaderInterface(ABC):
     def load_from_all(self, sources: tuple[str]) -> dict:
         """
         Look at all given sources and return an unified configuration.
-        :param sources: user inputs
+        :param sources: Source. Order is important. Last source will have precedence for the merge of configurations
         :return: configuration
         """