+++ src/wg_federation/data_transformation/loader/file/json_file_configuration_loader.py
@@ -1,19 +1,18 @@
+import json
 from typing import TextIO
-
-import yaml
 
 from wg_federation.data_transformation.loader.file.file_configuration_loader import FileConfigurationLoader
 
 
-class YamlFileConfigurationLoader(FileConfigurationLoader):
+class JsonFileConfigurationLoader(FileConfigurationLoader):
     """
-    Read any configuration from YAML files
+    Read any configuration from JSON files
     """
 
     @staticmethod
     def get_supported_source() -> str:
-        return 'yaml_file'
+        return 'json_file'
 
     @classmethod
     def _load_file(cls, file: TextIO) -> dict:
-        return yaml.safe_load(file)
+        return json.load(file)