+++ src/wg_federation/di/container.py
@@ -9,6 +9,7 @@ from wg_federation.constants import __version__
 from wg_federation.controller.configure_logging_controller import ConfigureLoggingController
 from wg_federation.controller.dispatcher.controller_dispatcher import ControllerDispatcher
 from wg_federation.data_transformation.loader.configuration_loader import ConfigurationLoader
+from wg_federation.data_transformation.loader.file.json_file_configuration_loader import JsonFileConfigurationLoader
 from wg_federation.data_transformation.loader.file.yaml_file_configuration_loader import YamlFileConfigurationLoader
 from wg_federation.input.manager.input_manager import InputManager
 from wg_federation.input.reader.argument_reader import ArgumentReader
@@ -75,7 +76,10 @@ class Container(containers.DynamicContainer):
         # input
         self.configuration_loader = providers.Singleton(
             ConfigurationLoader,
-            configuration_loaders=providers.List(providers.Singleton(YamlFileConfigurationLoader),),
+            configuration_loaders=providers.List(
+                providers.Singleton(YamlFileConfigurationLoader),
+                providers.Singleton(JsonFileConfigurationLoader),
+            ),
             logger=self.root_logger
         )